float debbuging, 0 overflow by 1

+++ float_form.c
@@ -24,12 +24,13 @@ char *f_form(va_list args, t_flags *st_flag)
 	{
 		n = va_arg(args, double);
 	}
-	else if (st_flag->fo_flag[2] == 1){
+	else if (st_flag->fo_flag[2] == 1)
+	{
 		n = va_arg(args, long double);
 	}
 	else
 	{
-		n = (float)va_arg(args, double);
+		n = va_arg(args, double);
 	}
 	//printf("prec in float is |%d|\n", st_flag->prec);
 	f_n = ft_ldtoa(n, st_flag->prec);
